Another bugfix for a special ghudson 'deleted' scenario.
When doing post-commit processing, it's not enough to notice that a
deleted item has a greater version number than its parent: this
assumes that the deleted item somehow belongs in the parent revision,
which is *not* necessarily true.  For example: {svn add newfile; svn
ci; svn rm newfile; svn ci}.  This will still leave a 'deleted' entry
for newfile, and during an update, our state-reporter will attempt to
report newfile as missing -- by deleting it from a txn where it never
existed. 

We debated adding new entry flags for this edge-case, but it's much
cleaner to allow our reporter->delete_path to allow no-op deletions.

* reporter.c (svn_repos_delete_path):  if the delete is a no-op, don't
  throw an error, just ignore.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@842518 13f79535-47bb-0310-9956-ffa450edef68