Fixing issue #797 in two stages.
This is the first half of the fix: store a new "version-url-rev"
wcprop at commit time, to compare with the entry committed rev
whenever a version-url is retrieved.  Also implement a new ra callback
function to retrieve the entry committed rev, but no callers yet.

The next step is to change ra_dav to do the comparison described
above.  It must validate its version-url wcprops by using the above
new callback, so they can never be out of sync with the entry again.
Whenever a version-url is retrieved, if the version-url-rev is present
but does not match the entry committed rev, then the version-url is
not valid, so we will erase it from the wcprop cache and return null.

* subversion/libsvn_ra_dav/ra_dav.h
  (SVN_RA_DAV__LP_VSN_URL_REV): Define new wcprop.

* subversion/libsvn_ra_dav/merge.c
  (bump_resource): Set mc->rev as the new wcprop.  Document this
  function.

* subversion/libsvn_wc/props.c 
  (svn_wc__wcprop_set): Handle null value correctly.  Remove
  dav-specific comment; if we're going to have it, it would better be
  at the interface documentation.
  (svn_wc__wcprop_get): Likewise remove dav-specific comment.

* subversion/include/svn_ra.h 
  (svn_ra_get_committed_rev_func_t): New function type.  This hurts me
  more than it hurts you.
  (svn_ra_callbacks_t): New callback `get_committed_rev'.

* subversion/libsvn_client/ra.c
  (set_wc_prop, get_wc_prop): Fix up comments.
  (get_committed_rev): New function.
  (svn_client__open_ra_session): Store get_committed_rev in the
  callbacks table, so ra_dav has a way to ask entry committed rev.

* subversion/include/svn_error_codes.h
  (SVN_ERR_BAD_REVISION): New error.

* subversion/include/svn_wc.h 
  (svn_wc_set_wc_prop, svn_wc_prop_set, svn_wc_prop_set):
  Documentation fixes.

* subversion/libsvn_wc/props.h 
  (svn_wc__wcprop_set): Documentation fix.

* subversion/libsvn_client/commit.c
  (get_ra_editor): Fix comment.

* subversion/libsvn_client/copy.c
  (repos_to_repos_copy): Fix comment.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@842629 13f79535-47bb-0310-9956-ffa450edef68