* www/project_status.html: Remove projected interim release.
The development heads of apr/apu and httpd have been too unstable the
last few days, so if we released an interim tarball, it would be the
first one that didn't obey the bootstrap rule of "you should update
immediately to head of svn, apr, apu, and httpd to test this".
Doesn't seem worth the effort.  I'd rather spend our time stabilizing
our dependencies and getting Subversion to match them.  If the
instability persists, then we'll have to start recommending dated
revisions of everything.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@842670 13f79535-47bb-0310-9956-ffa450edef68