Continue work on issue #730:  restartable checkouts.
It occurs to me that if set_path() or link_path() is called with
'start_empty' = TRUE, this not only means we should assume the
directory is empty of entries, but empty of props as well.  That's the
technical definition a 'revision' of a directory.  For an incomplete
wc directory, the client will need to report both the entries *and*
props it has.

* libsvn_repos/reporter.c (remove_directory_children): make this
  helper func remove all dir props as well as entries.

 (svn_repos_set_path, svn_repos_link_path): update comments in call to
  helper func.

* svn_ra.h (svn_ra_reporter_t):  update docstring for [set|link]_path.

* svn_repos.h (svn_repos_set_path, svn_repos_link_path): update docstrings.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@845991 13f79535-47bb-0310-9956-ffa450edef68