Fix 'svn diff URL URL' edge case.
If the changeset adds an empty file, no diff output was shown at all
for that event.  That's because the repos-diff editor was doing a diff
between two empty files.  But we still need to show the user that
*some* kind of add-event happened.  (And the 'patch' program still
interprets an empty diff as a signal to create an empty file.)

Note:  there's still a related ra_dav merge bug;  see issue #1347.

* libsvn_client/diff.c (diff_cmd_baton):  add new 'force_diff_output' flag.

  (diff_file_changed):  notice and honor this new flag.

  (diff_file_added): temporarily activate this flag before calling
  diff_file_changed, then deactivate it when finished.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@846223 13f79535-47bb-0310-9956-ffa450edef68