(Mostly) finish issue #1448 - svn:externals keeps connections open until end.  Since the demise of RA->close(), the only way to forcably shut down an RA session is to free the pool in which that session is allocated.  So, for all the sessions used by the externals code, we put those puppies in their own subpools and clear/destroy them when we're done with the session.
What's not been fixed is the fact that the "main" RA sessions for
checkout, update, and status (others?) aren't closed down until all
the externals handling is finished.

Oh, and I feel like I should credit Josh Pieper <jpieper@andrew.cmu.edu> 
for this change, even though I wrote every line of it.  He submitted a
patch as I was wrapping up this work which looked *almost identical*
to my own.  So, thanks to Josh for his effort, and for inadvertently
affirming the sense of my changes. :-)

* subversion/libsvn_client/externals.c
  (struct handle_external_item_change_baton): Declare the 'pool'
    member to be for short-lived scratchwork only.
  (handle_external_item_change): Clear the baton's pool before using
    it.  This will close previous RA sessions!
  (handle_externals_desc_change): Allocate and destroy a subpool for
    the hash diff mechanism.
  (svn_client__do_external_status): There are two nested loops here,
    so use two nested subpools.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@847386 13f79535-47bb-0310-9956-ffa450edef68