When setting a property, ensure that its name is compatible with RA-DAV.
This implementation restricts property names to the ASCII subset of an XML
Name.  A full Unicode XML Name would be a better thing to allow if we had a
convenient way of testing for it, but it is better to begin with a restricted
name space and to expand it later than to try the opposite.

* subversion/include/svn_error_codes.h
  (SVN_ERR_CLIENT_PROPERTY_NAME): New error.

* subversion/libsvn_client/prop_commands.c
  (is_valid_prop_name): New function.
  (svn_client_propset, svn_client_revprop_set): Validate the property name
    unless the property is being deleted.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@847602 13f79535-47bb-0310-9956-ffa450edef68