Command-line argument parsing errors: if a recognised but inappropriate option is found, just display an error message that says how to get help.
This is the same as what we do for an unrecognised option.  Don't display the
full help for the subcommand, because that is often long so that the original
command line and error message scroll off the screen, and it gets annoying
after the initial learning period, and it can encourage a habit of relying on
bad syntax (e.g. "svn add") to get help.

* subversion/clients/cmdline/main.c (main)
* subversion/svnadmin/main.c (main)
* subversion/svndumpfilter/main.c (main)
* subversion/svnlook/main.c (main)
  For an inappropriate option, display just a simple error message, not help.
  Fixed the description for "help help".
  Made quoting more consistent in other error messages.

* subversion/tests/clients/cmdline/getopt_tests_data/svn_help--version_stdout
  Updated accordingly.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@847620 13f79535-47bb-0310-9956-ffa450edef68