Prevent a potential race-condition in DAV commits. This race condition can happen when apache is experiencing heavy commit concurrency or if the network has long latency.
When proppatching svn:log, it's possible that the value of HEAD
changes in the small time-gap between fetching and checking out the
latest baseline, causing an out-of-dateness error.  This race is
specific to DeltaV semantics (and not those of SVN), so we hide this
race by allowing ra_dav to retry the operation a few times.

* libsvn_ra_dav/commit.c (apply_log_message): if apache tells us that
  a baseline is out-of-date when we try to check it out, retry fetching
  the latest baseline.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@848651 13f79535-47bb-0310-9956-ffa450edef68