Step 1 of N in removing most of cvs2svn material from the book.
This is part of a general shifting of the appendices.  We're removing
most of the content about cvs2svn and replacing it with pointers to
more up-to-date information, since cvs2svn is still changing.  The
pointers don't deserve their own appendix, so the current Appendix B,
which is dedicated to cvs2svn, will just go away, and the new
information (including mentions of the other converter tools) will go
in "Third Party Tools", with a ref from "Subversion For CVS Users".

Due to issue #1516, I can't do all these renames at once, hence there
will be several commits.

* doc/book/book/appb.xml: Removed.

* doc/book/book/ch05.xml: Anticipate new link to cvs2svn information.
    The new link is bogus until this change is finished, of course.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@848809 13f79535-47bb-0310-9956-ffa450edef68