Fix an obscure bug with updating deletions, and add a regression test for it.
If a directory has been deleted in the repository, and only a child of it has
been updated (and thereby deleted) in the WC, then a subsequent attempt to
update the directory in the WC would fail.

I discovered the anomaly in the source code; Greg Hudson analysed it,
determined the consequences, and devised a recipe to demonstrate it; I patched
the source code (changed one letter) and wrote the regression test.

* subversion/libsvn_repos/reporter.c
  (delta_dirs): Fix a typo responsible for a previously undiscovered bug.
  
* subversion/tests/clients/cmdline/update_tests.py
  (update_deletion_inside_out): New test.
  (test_list): Add the new test.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@849050 13f79535-47bb-0310-9956-ffa450edef68