Resolve issue #1920: Treat XML mime types as text, not binary.
Patch by: Branko Čibej <brane@xbc.nu>

* subversion/libsvn_subr/validate.c
  (svn_mime_type_is_binary): Return false if the mime-type ends in
    "+xml" or "/xml".

There was an alternate patch by Kalle Olavi Niemitalo <kon@iki.fi> in
http://subversion.tigris.org/servlets/ReadMsg?list=dev&msgNo=66354,
which is a little more robust because it deals with charset addenda to
mime-types.  There was also a general objection by Ben Reser, in
http://subversion.tigris.org/servlets/ReadMsg?list=dev&msgNo=68882,
where he points out that the whole approach is wrong, and that this
should be configurable.  I agree (as probably most developers would),
but it's always been understood that svn_mime_type_is_binary() is
basically a kluge to get the common cases right until such time as we
have a better system.  Incrementally improving what we have, based on
user feedback, is still a good idea IMHO.

See the threads linked to from the issue.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@850174 13f79535-47bb-0310-9956-ffa450edef68